autoboot: Make all prompts configurable
This is a multi-purpose commit. Though it is hard to split into discrete
changes.

1. Allows the non-keyed prompt to be configured

This is self-explanatory. This allows better customization for the
integrator. Though, more to the point, this reduces the confusion that
comes from the option name and description. Now it is used for all
auto-boot prompts, not only for the keyed prompt.

2. Redraws using ANSI escapes

This is required for (1), as we can't backspace over the arbitrary
amount of characters to redraw the countdown.

This is done through resetting the column to 1 and clearing the line for
maximum compatibility. Tested against serial, default dm_video and the
truetype console.

Signed-off-by: Samuel Dionne-Riel <samuel@dionne-riel.com>