treewide: Identify as Tow-Boot
The idea is not to hide this is built on top of U-Boot.

The goal of this change is to **clearly identify this is not a standard
U-Boot build**.

This way, support questions shouldn't end up in U-Boot's support
channels. Hopefully.

What about `spl_phase_name`??? Well... Remove one location where U-Boot
identifier would be used front and center... And really, it'll help
identify which part of the board init is happening! Win-win!