Add tb_menu command
This command allows creating *imperatively* dynamic menus. This is done
through building the menus with commands, rather than through hardly
dynamic numbered environment variables.

This uses PDCurses for *more correct* TUI interfaces. Though the
approach could be used without; changing the existing bootmenu to an
imperative approach could be done too.

Using PDCurses, though, allows for more complex interactions to happen
in a more standard way. While curses are named appropriately, from the
cursing of developers using it, it had undeniable advantages compared
to manually doing the hard work of building text-based interfaces.

This first draft is sufficient for the desired goals. This is probably
not the final form of the menu.