video: Add video_sync_dirty
This is used to work around the slow redraw of a full video console
display.

The `video_sync_dirty` function is *supposed* to be called in situations
where it makes sense for a dynamic user use case. The easiest example
would be in the sibling patch, where it is applied to `tstc` and `getc`
methods of graphical input methods.

This is not supposed to go upstream. This is a workaround for the issue,
which would be better fixed with some kind of co-operative "multi"
tasking, where instead the video-uclass would add a task to sync
"whenever the system can" (if not asked for already).