button-sun4i-lradc: Provide UCLASS_BUTTON driver
This is based on the hacky driver by megi:

 - https://megous.com/git/u-boot/commit/?h=opi-v2020.01&id=0ab6225154c3d8b74f06fb3b181b52a9a64b4602

The main change here is it provides a UCLASS_BUTTON driver.

The driver is still not upstreamable in its current form (I think), but
at least the behaviour should be consistent with other UCLASS_BUTTON
drivers, meaning the generic usage should continue working once a better
driver is written-up.