[Tow-Boot] rockchip: Make Rockchip download mode in SPL configurable
This "maskrom button" handling isn't as useful as it seems, first, since
it requires a working SPL, and given our usual workflow, the working SPL
*should* imply a working U-Boot proper.

On some devices, the Pinephone Pro, this is conflicting with the better
semantics to allow USB Mass Storage vs. External boot by holding the
volume buttons.