[HACK] implement `do_poweroff` for rk8xx
In reality this should probably be moved into a "shim" library like
`drivers/power/pmic-poweroff.c`.

Though it's unclear if it would be appropriate, considering that would
not take `rockchip,system-power-controller` into account.

In actuality it will require someone that better knows PMIC, PSCI,
for RK3399 and other devices, to know how to do the correct
implementation.

For the time being, this works.