power: pmic: rk8xx: Workaround pmic failure when probed before relocation
When the sysreset is added as child of the pmic the pmic is probed
before relocation. That probe fails, and subsequent attempts to probe
after reloaction fail as well.

As a workaround do not bind the sysreset before relocation.

Signed-off-by: Michal Suchanek <msuchanek@suse.de>
Origin: https://patchwork.ozlabs.org/project/uboot/patch/20220805113219.7316-1-msuchanek@suse.de/