smbios: identify as Tow-Boot and expose full version string
The SMBIOS spec states for BIOS Version:

> String number of the BIOS Version. This value is a free-form string
> that may contain Core and OEM version information.

I think this is the perfect string to use here.

Changing the vendor here is not meant to fake that we're not U-Boot,
but means that we correctly signal that the behaviour may differ from
upstream U-Boot. This is important, we don't want users going to the
exposed vendor (e.g. U-Boot) and tying up their ressources on issues
of a forked project.