pinephone-pro: Perform PMIC setup on boot (increase input current limit)
The limit needs to be raised to prevent bootloops on empty battery.

(This is a bit too late for battery-less boot, it would need to be done
in TPL very early on, for that) This is still slightly better than nothing,
though.

Signed-off-by: Ondrej Jirman <megi@xff.cz>
Origin: https://xff.cz/git/u-boot/commit/?h=ppp&id=7f8238fd608290152b143322178a5be21a447dc1