[Tow-Boot] efi_loader: Disable EFI reset for RK3399
While we have added poweroff support in Tow-Boot for rk8xx, it is broken
once the operating system takes over.

Poweroff from EFI poweroff will hang the system. This is sometimes
tricky to detect since when it happens, LEDs and most other things are
turned off. Meaning that the system *appears* off, but is actually still
using power.

For now, pretend RK3399 can't do it (since it can't really). Assuming
Linux is the running Operating System, it will fall back to non-EFI
methods of doing a poweroff.

Signed-off-by: Samuel Dionne-Riel <samuel@dionne-riel.com>