drivers: video: Clear framebuffer when reserving
This fixes an issue where in some situations, the fixed framebuffer is
re-used (understandably), and with a warm reset is shown again briefly
during video initialization, until U-Boot renders something new.

This can be observed on the Pinebook Pro (RK3399, eDP), once the reset
issue of the panels are worked around correctly.

Signed-off-by: Samuel Dionne-Riel <samuel@dionne-riel.com>